+++ CHANGELOG.md
@@ -21,6 +21,11 @@ Current
 - [Upgrades to netty 4.1.42.45.Final to address CVE-2019-20444 and CVE-2019-20445](https://github.com/yahoo/fili/pull/1006)
 
 ### Added:
+- [Add parameters for output logical metric info to core classes in the ProtocolMetric API](https://github.com/yahoo/fili/pull/1020)
+    * Add `outputMetadata` parameter to the `ProtocolMetric` and `MetricTransformer` interfaces
+        - Allows for a consistent way to name and track the result metric transformation, instead of deferring the renaming
+          responsibility to the implementations. 
+
 - [Add COUNT(\*) support in fili-sql](https://github.com/yahoo/fili/pull/992)
    * When there is a `count` metric that uses `countMaker`, it will be translated into a COUNT(\*) in SQL query.